fix routing toarticle and category pages

+++ src/components/CommentCard.jsx
@@ -1,11 +1,13 @@
 import { deleteCommentById } from "../../api";
+import { currentUser } from "../App";
 export default function CommentCard({ comment }) {
   const { comment_id, body, article_id, author, votes, created_at } = comment;
+  const userName = currentUser._currentValue.userName;
 
   function handleClick_DeleteComment() {
-    deleteCommentById(comment_id);
+    if (userName === author) deleteCommentById(comment_id);
   }
-  console.log();
+
   return (
     <section className="comment-card">
       <div>
@@ -16,9 +18,11 @@ export default function CommentCard({ comment }) {
         <div>{`${created_at.slice(0, 10)}`}</div>
       </div>
       <div>
-        <button onClick={handleClick_DeleteComment} type="button">
-          Delete comment
-        </button>
+        {userName === author && (
+          <button onClick={handleClick_DeleteComment} type="button">
+            Delete comment
+          </button>
+        )}
       </div>
     </section>
   );
